backend/requirements.txt
fastapi
uvicorn
pymupdf
langchain
langchain-huggingface
faiss-cpu
langchain-community
mysql-connector-python
sqlalchemy 
pymysql